allows these objects to be serialized
Provides a mechanism for the activity to save the state of the permissions
being requested of the user. The activity can be closed, restarted, and
resumed. Without saving the state of the permissions being requested, the
activity has no way to know if it is already in the process of requesting
the user for a permission.